поправка с обработкой ошибки/ результат int

+++ app/schemas/result.py
@@ -2,6 +2,6 @@ from pydantic import BaseModel
 
 
 class MatchResult(BaseModel):
-    match_percentage: float
+    match_percentage: int
     matched: list[str]
     didnt_match: list[str]
